+++ lutra-ir/src/printer/test.rs
@@ -12,36 +12,36 @@ fn print_01() {
     let main =
       let 1 = (
         func 2 -> [
-            fn.2+0: float,
-            fn.2+0: float,
-            fn.2+0: float
-        ]: [float]
-      ): func (float) -> [float];
+            fn.2+0: float64,
+            fn.2+0: float64,
+            fn.2+0: float64
+        ]: [float64]
+      ): func (float64) -> [float64];
     
-      let 2 = var.1: func (float) -> [float];
+      let 2 = var.1: func (float64) -> [float64];
       {
         (
-            call var.2: func (float) -> [float],
-            3.5: float
-        ): [float],
+            call var.2: func (float64) -> [float64],
+            3.5: float64
+        ): [float64],
         (call
             (
                 func 3 -> [
-                    fn.3+0: int,
-                    fn.3+1: int,
-                ]: [int]
-            ): func (int, int) -> [int],
-            6: int,
-            7: int,
-        ): [int],
+                    fn.3+0: int64,
+                    fn.3+1: int64,
+                ]: [int64]
+            ): func (int64, int64) -> [int64],
+            6: int64,
+            7: int64,
+        ): [int64],
         (call
-            external.0: func (int, int) -> int,
-            6: int,
-            2: int
-        ): int
-      }: {[float], [int], int}
-      .1:[int]
-      .[0]:int
+            external.0: func (int64, int64) -> int64,
+            6: int64,
+            2: int64
+        ): int64
+      }: {[float64], [int64], int64}
+      .1:[int64]
+      .[0]:int64
     "#), @r#"
     let externals = [
       std_int_add,
@@ -50,34 +50,34 @@ fn print_01() {
     let main =
       let 1 = (
         func 2 -> [
-          fn.2+0: float,
-          fn.2+0: float,
-          fn.2+0: float,
-        ]: [float]
-      ): func (float) -> [float];
-      let 2 = var.1: func (float) -> [float];
+          fn.2+0: float64,
+          fn.2+0: float64,
+          fn.2+0: float64,
+        ]: [float64]
+      ): func (float64) -> [float64];
+      let 2 = var.1: func (float64) -> [float64];
       {
         (
-          call var.2: func (float) -> [float], 
-          3.5: float, 
-        ): [float],
+          call var.2: func (float64) -> [float64], 
+          3.5: float64, 
+        ): [float64],
         (
           call (
             func 3 -> [
-              fn.3+0: int,
-              fn.3+1: int,
-            ]: [int]
-          ): func (int, int) -> [int], 
-          6: int, 
-          7: int, 
-        ): [int],
+              fn.3+0: int64,
+              fn.3+1: int64,
+            ]: [int64]
+          ): func (int64, int64) -> [int64], 
+          6: int64, 
+          7: int64, 
+        ): [int64],
         (
-          call external.0: func (int, int) -> int, 
-          6: int, 
-          2: int, 
-        ): int,
-      }: {[float], [int], int}
-      .1: [int]
-      .[0]: int: int
+          call external.0: func (int64, int64) -> int64, 
+          6: int64, 
+          2: int64, 
+        ): int64,
+      }: {[float64], [int64], int64}
+      .1: [int64]
+      .[0]: int64: int64
     "#);
 }
